backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
firebase-admin==6.4.0
PyJWT==2.8.0
python-dotenv==1.0.0
bcrypt==4.1.2
pandas==2.1.4
scikit-learn==1.3.2
networkx==3.2.1
openpyxl==3.1.2
joblib==1.3.2
pydantic==2.5.0
numpy==1.24.3
requests==2.31.0